+++ test/spec/xform.spec.js
@@ -105,13 +105,13 @@ describe( 'XForm', () => {
 
         it( 'outputs errors for calculations without form control that refer to external ' +
             'clinicaldata instance but do not have the oc:external="clinicaldata" bind', () => {
-            expect( arrContains( result1.errors, /refers to external clinicaldata without the required "external" attribute/i ) ).to.equal( true );
-        } );
+                expect( arrContains( result1.errors, /refers to external clinicaldata without the required "external" attribute/i ) ).to.equal( true );
+            } );
 
         it( 'outputs errors for binds with oc:external="clinicaldata" that do not ' +
             'do not have a calculation that refers to instance(\'clinicaldata\')', () => {
-            expect( arrContains( result1.errors, /not .* calculation referring to instance\('clinicaldata'\)/i ) ).to.equal( true );
-        } );
+                expect( arrContains( result1.errors, /not .* calculation referring to instance\('clinicaldata'\)/i ) ).to.equal( true );
+            } );
     } );
 
     describe( 'with incorrect appearance usage', () => {
@@ -172,8 +172,10 @@ describe( 'XForm', () => {
         const result = validator.validate( xf );
 
         it( 'outputs warnings', () => {
-            expect( result.warnings.length ).to.equal( 1 );
-            expect( arrContains( result.warnings, /Duplicate nodenames found./i ) ).to.equal( true );
+            expect( result.warnings.length ).to.equal( 2 );
+            expect( arrContains( result.warnings, /Duplicate .* name "a" found/i ) ).to.equal( true );
+            expect( arrContains( result.warnings, /Duplicate .* name "g" found/i ) ).to.equal( true );
+            expect( arrContains( result.warnings, /Duplicate .* name "b" found/i ) ).to.equal( false );
         } );
 
     } );
@@ -191,11 +193,11 @@ describe( 'XForm', () => {
     } );
 } );
 
-describe('XForm Class', () => {
+describe( 'XForm Class', () => {
     it( 'should throw if XForm string not provided', () => {
         let failure = () => { new XForm(); };
         expect( failure ).to.throw();
-    });
+    } );
 
     describe( 'nsPrefixResolver method', () => {
         const xf = new XForm( loadXForm( 'model-only.xml' ) );
@@ -211,8 +213,8 @@ describe('XForm Class', () => {
         const xf = new XForm( loadXForm( 'model-only.xml' ) );
         it( 'should parse model if it wasn\'t parsed already', () => {
             expect( typeof xf.model === 'undefined' ).to.equal( true );
-            xf.enketoEvaluate('floor(1)');
+            xf.enketoEvaluate( 'floor(1)' );
             expect( typeof xf.model === 'undefined' ).to.equal( false );
         } );
     } );
-});
+} );